// src/app/models/appointment.model.ts
export class Appointment {
  Id: number;
  userId: string;
  appointmentDepartmentId: number;
  appointmentStatusId: number;
  appointmentTitle: string;
  appointmentStartDate: string;
  appointmentEndDate: string;
  appointmentDescription: string;
  appointmentBGColor: string;
  CreatedBy: string;
  CreationDate: string;
  UpdatedBy: string;
  UpdationDate: string;
  Description: string;
}